+++ CHANGELOG.md
@@ -7,6 +7,8 @@
         * `Klock`: `2.0.0` -> `2.0.1`
         * `Ktor`: `1.4.2` -> `1.4.3`
         * `MicroUtils`: `0.4.6` -> `0.4.9`
+* `API Extensions`:
+    * New function `buildBot`
 
 ## 0.30.9